marketselect: relocate Function 1 into geolift_helpers/marketselect/helpers
Faithful-first naming: the market-selection port lives under the GeoLift
tree (geolift_helpers/marketselect/helpers/) rather than a neutral
placeholder, matching the one-estimator-one-package convention with
marketselect as the method subpackage. No code change to the similarity
leaves; test import updated to the new path.
