+++ mlsynth/utils/geolift_helpers/__init__.py
@@ -1,0 +1,7 @@
+"""Helpers for the GeoLift port.
+
+GeoLift is an end-to-end geo-experiment methodology (market selection ->
+power analysis -> synthetic-control estimation). This package collects the
+faithful port of its pieces as method subpackages — starting with
+``marketselect`` (candidate market nomination and ranking).
+"""
