+++ mlsynth/utils/geolift_helpers/marketselect/__init__.py
@@ -1,0 +1,6 @@
+"""GeoLift market selection.
+
+Nominates candidate test-market sets and scores them. ``helpers`` holds the
+treatment-agnostic, estimator-independent leaves (similarity ranking, candidate
+generation) that the selection pipeline composes.
+"""
